Implement TextProcessorService microservice and finish project.

+++ src/main/java/com/algaworks/post_service/post/service/api/model/output/PostOutput.java
@@ -15,6 +15,6 @@ public class PostOutput {
     private String title;
     private String body;
     private String author;
-    private Integer wordCount;
+    private Long wordCount;
     private BigDecimal calculatedValue;
 }
